make tests more re-usable
