+++ tests/architecture_prototype/test_graph.py
@@ -1,7 +1,5 @@
 import copy
 import json
-import shutil
-from pathlib import Path
 from typing import Any, Dict, Text
 
 from rasa.architecture_prototype import graph
@@ -12,6 +10,7 @@ from rasa.shared.core.constants import ACTION_LISTEN_NAME
 from rasa.shared.core.events import ActionExecuted, UserUttered
 from rasa.shared.core.trackers import DialogueStateTracker
 from rasa.shared.nlu.constants import INTENT_NAME_KEY
+from tests.architecture_prototype import conftest
 from tests.architecture_prototype.graph_schema import (
     full_model_train_graph_schema,
     predict_graph_schema,
@@ -21,15 +20,8 @@ import rasa.nlu.registry
 import rasa.core.registry
 
 
-def clean_directory():
-    # clean up before testing persistence
-    cache_dir = Path("model")
-    shutil.rmtree(cache_dir, ignore_errors=True)
-    cache_dir.mkdir()
-
-
 def test_train_nlu():
-    clean_directory()
+    conftest.clean_directory()
     graph.fill_defaults(nlu_train_graph_schema)
     serialized = json.dumps(nlu_train_graph_schema)
     deserialized = json.loads(serialized)
@@ -41,43 +33,16 @@ def test_train_nlu():
     print(trained_components)
 
 
-def test_train_full_model():
-    clean_directory()
-    graph.fill_defaults(full_model_train_graph_schema)
-    graph.visualise_as_dask_graph(full_model_train_graph_schema, "full_train_graph.png")
-    core_targets = ["train_memoization_policy", "train_ted_policy", "train_rule_policy"]
-    nlu_targets = [
-        "train_classifier",
-        "train_response_selector",
-        "train_synonym_mapper",
-    ]
-    trained_components = graph.run_as_dask_graph(
-        full_model_train_graph_schema, core_targets + nlu_targets
-    )
-    print(trained_components)
+def test_train_full_model(trained_model: Dict):
+    print(trained_model)
 
 
-def test_train_load_predict():
-    clean_directory()
-
-    graph.fill_defaults(full_model_train_graph_schema)
-    graph.visualise_as_dask_graph(full_model_train_graph_schema, "full_train_graph.png")
-    core_targets = ["train_memoization_policy", "train_ted_policy", "train_rule_policy"]
-    nlu_targets = [
-        "train_classifier",
-        "train_response_selector",
-        "train_synonym_mapper",
-    ]
-    graph.run_as_dask_graph(full_model_train_graph_schema, core_targets + nlu_targets)
-
-    graph.fill_defaults(predict_graph_schema)
-    graph.visualise_as_dask_graph(predict_graph_schema, "predict_graph.png")
+def test_train_load_predict(prediction_graph: Dict[Text, Any]):
+    graph.visualise_as_dask_graph(prediction_graph, "predict_graph.png")
 
     predictions = graph.run_as_dask_graph(predict_graph_schema, ["select_prediction"])
     for prediction in predictions.values():
         print(prediction)
-
-    # TODO: Fix empty message
 
     # TODO: should we be saving model_metadata?
     # TODO: stuff like rasa.utils.train_utils.update_deprecated_loss_type changing wont force re-train?
@@ -132,10 +97,8 @@ def test_serialize_graph_schema():
     graph.run_as_dask_graph(deserialized, core_targets + nlu_targets)
 
 
-def test_model_prediction_with_and_without_nlu():
-    graph.fill_defaults(predict_graph_schema)
-
-    model = Model(predict_graph_schema)
+def test_model_prediction_with_and_without_nlu(prediction_graph: Dict[Text, Any]):
+    model = Model(prediction_graph)
 
     prediction_with_nlu = model.handle_message(
         DialogueStateTracker.from_events(
@@ -143,7 +106,6 @@ def test_model_prediction_with_and_without_nlu():
         ),
         message=UserMessage(text="hello"),
     )
-    print(prediction_with_nlu)
 
     prediction_without_nlu = model.predict_next_action(
         DialogueStateTracker.from_events(
@@ -154,6 +116,5 @@ def test_model_prediction_with_and_without_nlu():
             ],
         )
     )
-    print(prediction_without_nlu)
 
     assert prediction_with_nlu == prediction_without_nlu
